make paths relative to CWD, after all
the test suite actually relies on it. it would be possible to adjust it,
but there is not much reason to make paths relative to HOME (as we
support convenient tilde expansion). so use the least invasive approach,
which is simply the old behavior. adjust the documentation accordingly.

This reverts commit c07eb05e6121f77eeef01092fdbe560c2ba42c8e.
(cherry picked from commit 15d57b95b7c72acac55c22e0c70771a7565a126f)