*** implement fallback path in UIDVALIDITY recovery
mailboxes with mostly messages with no Message-Id (such as Drafts and
Templates) cannot recover the usual way.

this patch implements a statistical method. as such, it has a (very)
slight chance to yield a false positive which would subsequently clobber
actual messages (specifically, the ones which were not sufficient to
throw off the algorithm). it would be possible to be even stricter, but
then any actual modification of the box would make the algorithm fail.
and even in this case, it would be still possible that even though the
UIDs match (and thus nothing gets deleted), the actual messages are in
reality different (though this is ridiculously unlikely).
the added tolerance doesn't help with stores which are synced seldomly,
as too much is likely to happen in the mean time.

*** actually make it stricter?