*** workaround exchange being unhapply about messages with binary content
*** this is is an *untested* rebase; originally submitted against 1.3

Those where bogus messages with the raw attachment in binary but with
base 64 headers correctly set.
Near 100% (if not 100%) of those where in the sent folder and are
probably the result of gmail + buggy email client (but you can still
open the attachment with gmail !)

this is a possible fix for https://sourceforge.net/p/isync/bugs/22/ and
a lot of related reports.

patch by Florian Lombard <f.lombard@montmirail.com>:

============

the treatment should be the same as for messages with excesively long
lines - MIME-encoding (presumably as quoted-printable).