*** introduce NotifierCmd
*** actually implement most of it

this could take advantage of wip/better-stderr.

to be actually useful, this would need to report both internally
generated and detected events, but typically only for one side.
this is contrary to how a natural implementation would work, which
would report detected events as internal events for the opposite side,
but of course only if a corresponding sync action was actually executed.

=> a directory watcher seems more appropriate in the end