automate setting package version
this avoids the need for bumping the version, which is particularly
helpful if one doesn't know yet whether the next release will be a
patch, minor, or major.

we cache the version extracted from git, which also provides a fallback
for the case of somebody rebuilding configure from a tar-ball.

note that it's impossible to determine the version at configure time, so
after git-tagging you need to remember to run version.sh (or autoconf)
prior to rolling a tar-ball.